pyproject.toml
[tool.poetry]
name = "langserve-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
rag-conversation = {path = "packages/rag-conversation", develop = true}
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.1"
faiss-cpu = "^1.8.0"
pymupdf = "^1.24.0"
huggingface-hub = "^0.22.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
